+++ src/Bootstrap3Jumbotron.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types=1);
-/**
- * Created by PhpStorm.
- * User: g3d
- * Date: 16/05/2017
- * Time: 07:52
- */