+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
-Full changelog for ZF3 view helpers
+Full changelog for ZF3 view helpers library.
+
+## Initial release - 2017-05-16
+
+* Added Bootstrap 4 Jumbotron view helper.